example/requirements.txt
pytz
Django
djangorestframework
sspdatatables
django-countries